@@ -1,28 +1,53 @@
 {
-  "description": "NativeScript Application",
-  "license": "SEE LICENSE IN <your-license-filename>",
-  "readme": "NativeScript Application",
-  "repository": "<fill-your-repository-here>",
   "nativescript": {
-    "id": "org.nativescript.flashlightdemo",
-    "tns-ios": {
-      "version": "2.5.0"
-    },
+    "id": "org.nativescript.demo",
     "tns-android": {
-      "version": "2.5.0"
+      "version": "3.1.1"
+    },
+    "tns-ios": {
+      "version": "3.1.0"
     }
   },
   "dependencies": {
-    "nativescript-flashlight": "file:../../nativescript-flashlight",
-    "tns-core-modules": "2.5.1"
+    "nativescript-flashlight": "../src",
+    "nativescript-theme-core": "^1.0.4",
+    "nativescript-unit-test-runner": "^0.3.4",
+    "tns-core-modules": "^3.1.1"
   },
   "devDependencies": {
-    "babel-traverse": "6.16.0",
-    "babel-types": "6.16.0",
-    "babylon": "6.11.4",
+    "awesome-typescript-loader": "~3.1.3",
+    "babel-traverse": "6.12.0",
+    "babel-types": "6.11.1",
+    "babylon": "6.8.4",
+    "copy-webpack-plugin": "~4.0.1",
+    "extract-text-webpack-plugin": "~3.0.0",
+    "filewalker": "0.1.2",
+    "jasmine-core": "^2.5.2",
+    "karma": "^1.3.0",
+    "karma-jasmine": "^1.0.2",
+    "karma-nativescript-launcher": "^0.4.0",
     "lazy": "1.0.11",
-    "nativescript-dev-typescript": "^0.3.5",
-    "tns-platform-declarations": "2.5.1",
-    "typescript": "~2.1.6"
+    "nativescript-css-loader": "~0.26.0",
+    "nativescript-dev-typescript": "libs",
+    "nativescript-dev-webpack": "^0.7.3",
+    "raw-loader": "~0.5.1",
+    "resolve-url-loader": "~2.1.0",
+    "tns-platform-declarations": "^3.1.0",
+    "tslint": "~5.4.3",
+    "typescript": "~2.3.0",
+    "webpack": "~3.2.0",
+    "webpack-bundle-analyzer": "^2.8.2",
+    "webpack-sources": "~1.0.1"
+  },
+  "scripts": {
+    "build.plugin": "cd ../src && npm run build",
+    "ci.tslint": "npm i && tslint --config '../tslint.json' 'app/**/*.ts' --exclude '**/node_modules/**'",
+    "ns-bundle": "ns-bundle",
+    "publish-ios-bundle": "npm run ns-bundle --ios --publish-app",
+    "generate-android-snapshot": "generate-android-snapshot --targetArchs arm,arm64,ia32 --install",
+    "start-android-bundle": "npm run ns-bundle --android --run-app",
+    "start-ios-bundle": "npm run ns-bundle --ios --run-app",
+    "build-android-bundle": "npm run ns-bundle --android --build-app",
+    "build-ios-bundle": "npm run ns-bundle --ios --build-app"
   }
 }
