@@ -1,2 +1,3 @@
-﻿import application = require("application");
+﻿import "./bundle-config";
+import * as application from 'tns-core-modules/application';
 application.start({ moduleName: "main-page" });
