@@ -1,4 +1,5 @@
-export declare abstract class FlashLightCommon {
+import { Observable } from 'tns-core-modules/data/observable';
+export declare abstract class Common extends Observable {
     isOn: boolean;
     toggle(arg: any): void;
     protected checkAvailability(): void;
